Get rid of *keywordize*; give a 'keywordize' parameter to 'decode'.
Previously we'd get issues after setting *keywordize* to false and
then calling e.g. `(map decode ...)`; dynamic variables are inherited
from the calling context and the calling context in that case is 'map',
so we'd get the default.

Using dynamic variables at all is kind of crazy; this is much nicer!

(All tests pass).