Move dependency specifications to deps.edn
This enables libraries and applications that use tools.deps to retrieve
this library directly from its GitHub repository, as code, rather than
from an artifact repository such as Clojars. This has two main benefits:
(a) it improves security by reducing opportunities for malicious code
injection via hijacking of the artifact repository or the artifact
publication keys; (b) it enables projects using tools.deps to easily use
any version of the library that exists in the GitHub repository — any
commit on any branch — even forks — and they don’t have to wait for the
maintainers to cut and publish a new release. (This would hopefully also
reduce pressure on the maintainers.)

Using plugin lein-tools-deps to “forward” (more or less) those
dependency specifications to Leiningen.

Motivation for using lein-tools-deps rather than switching to tools.deps
altogether: so we can continue to use Leiningen, so that the maintainers
of this library don’t have to change their workflows and we don’t have
to change our CI configuration. This is an attempt to hew to the ethos
that it’s best to change one thing at a time; as it appears that we
*can* change how we specify our dependencies without changing our build
tool, then we *should* do so.

For more discussion of this change, see this GitHub Pull Request:
  https://github.com/clj-commons/clj-yaml/pull/4

Docs links:
  * deps.edn: https://clojure.org/reference/deps_and_cli#_deps_edn
  * lein-tools-deps: https://github.com/RickMoynihan/lein-tools-deps