Unbreak YAMLCodec's decode
Version 0.7.169 of clj-yaml changed YAMLCodec's decode function
signature. We assumed this was an internal-only function but have since
learned that this is sometimes used in the wild.

This change breaks 0.7.169 but unbreaks prior versions.

- old sig: (decode [data keywords])
- 0.7.169: (decode [data keywords unknown-tag-fn])
- new sig: (decode [data opts])

We interpret opts as an options map, if it is not of type map we assume
legacy support for old sig and interpret it as keywords boolean.

Also:
- added a note to user guide about clj-yaml's keyword args fns
- made high level api fn signatures that use keyword args consistent

Closes #67