Address new requirements for vulnerability scan
OSS Index now requires credentials.

Since clj-watson can specify these new credentials without change,
switch from nvd-clojure to clj-watson.

Turf old `nvd_helper_project` dir in favour of terser `nvd-scan` dir.

We separated out nvd scanning to a subdir on the strong advice from
nvd-clojure docs. There is no such recommendation for this for
clj-watson, but we keep the subdir because it helps with CI caching of
the vulnerability database and invalidating that cache only when necessary.

For #165